requirements.txt
deepdiff==5.5.0
PyYAML==5.3.1
rich
psutil